Fix hang after SIGINT in debug mode
Sending SIGINT (^c) to the process when in debug mode causes the
process to (at least appear to) hang, because the serial port is still
open.  Instead of deleting our reference to the serialport, call its
close method, and then call process.exit in the handler for the close
event.